sunday
finally 2 major bugs fixed in addtocart component

+++ frontend/src/components/User/Customer/AddToCart.js
@@ -10,10 +10,11 @@ const AddToCart = () => {
   const [date, setDate] = useState(new Date());
   // const [cartItems, setCartItems] = useState(null);
   const [selectedProduct, setSelectedProduct] = useState(null);
-  const [quantity, setQuantity] = useState(null);
+  const [quantity, setQuantity] = useState([]);
   const [trigger, setTrigger] = useState(false);
   const [total, setTotal] = useState(0);
   const [index, setIndex] = useState(null);
+  const [deleteItem, setDeleteItem] = useState(false);
   const [username, setUsername] = useState(null);
   const [boughtItems, setBoughtItems] = useState([]);
   const dispatch = useDispatch();
@@ -38,9 +39,13 @@ const AddToCart = () => {
   }, [getUserTrigger]); */
 
   useEffect(() => {
-    setQuantity(Array(cartItems.length).fill(1));
-    totalCounter();
-  }, []);
+    if (deleteItem) {
+    } else {
+      //setQuantity(Array(cartItems.length).fill(1));
+      setQuantity((prev) => [...prev, 1]);
+      totalCounter();
+    }
+  }, [cartItems]);
 
   /*useEffect(() => {
     if (selectedProduct) {
@@ -53,11 +58,11 @@ const AddToCart = () => {
     }
   }, [selectedProduct]);*/
 
-  /*useEffect(() => {
+  useEffect(() => {
     if (boughtItems) {
       handleBuyNow();
     }
-  }, [boughtItems]);*/
+  }, [boughtItems]);
 
   function totalCounter() {
     let total = 0;
@@ -70,12 +75,12 @@ const AddToCart = () => {
     setTotal(total);
   }
 
-  /*function arraySort() {
+  function handleBoughtProducts() {
     const newArray = cartItems.map((item) => {
       return item._id;
     });
     setBoughtItems(newArray);
-  }*/
+  }
 
   function handleBuyNow() {
     axios
@@ -93,9 +98,16 @@ const AddToCart = () => {
   }
 
   useEffect(() => {
-    dispatch(removeCartItem(index));
+    if (deleteItem) {
+      let newArray = [...quantity];
+      newArray.splice(index, 1);
+      setQuantity(newArray);
+      dispatch(removeCartItem(index));
+      setDeleteItem(false);
+      setIndex(null);
+    }
   }, [index]);
-  console.log(cartItems);
+
   return (
     <div
       className={
@@ -125,106 +137,104 @@ const AddToCart = () => {
         <h1 className="ml-4 text-2xl">Cart</h1>
       </div>
       <div className="h-[90%] border-b-2 border-gray-300 border-opacity-20 pb-2 overflow-scroll pt-2 ">
-        {cartItems &&
-          cartItems.map((item, index) => {
-            return (
-              <div className="h-[15%] bg-gray-300 bg-opacity-50 flex items-center pt-2 pb-2 pl-2 mt-1 relative">
-                <button
-                  className="absolute left-1 top-1 text-white"
-                  onClick={() => {
-                    setIndex(index);
-                  }}
+        {cartItems.map((item, index) => {
+          return (
+            <div className="h-[15%] bg-gray-300 bg-opacity-50 flex items-center pt-2 pb-2 pl-2 mt-1 relative">
+              <button
+                className="absolute left-1 top-1 text-white"
+                onClick={() => {
+                  setIndex(index);
+                  setDeleteItem(true);
+                }}
+              >
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 24 24"
+                  fill="currentColor"
+                  class="w-6 h-6"
                 >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    viewBox="0 0 24 24"
-                    fill="currentColor"
-                    class="w-6 h-6"
-                  >
-                    <path
-                      fill-rule="evenodd"
-                      d="M5.47 5.47a.75.75 0 011.06 0L12 10.94l5.47-5.47a.75.75 0 111.06 1.06L13.06 12l5.47 5.47a.75.75 0 11-1.06 1.06L12 13.06l-5.47 5.47a.75.75 0 01-1.06-1.06L10.94 12 5.47 6.53a.75.75 0 010-1.06z"
-                      clip-rule="evenodd"
-                    />
-                  </svg>
-                </button>
-                <img
-                  src={item.productPicture[0]}
-                  className="w-[35%] h-full object-cover rounded-sm"
-                ></img>
-                <div className="ml-2">
-                  <h1 className="text-2xl ">{item.productName}</h1>
-                  <h2 className="mt-2">{item.productNewPrice}$</h2>
-                </div>
-                <div className="h-full absolute right-2 ">
-                  <div className=" h-full fl2 ">
-                    <button>
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        viewBox="0 0 24 24"
-                        fill="currentColor"
-                        class="w-5 h-5"
-                        onClick={() => {
+                  <path
+                    fill-rule="evenodd"
+                    d="M5.47 5.47a.75.75 0 011.06 0L12 10.94l5.47-5.47a.75.75 0 111.06 1.06L13.06 12l5.47 5.47a.75.75 0 11-1.06 1.06L12 13.06l-5.47 5.47a.75.75 0 01-1.06-1.06L10.94 12 5.47 6.53a.75.75 0 010-1.06z"
+                    clip-rule="evenodd"
+                  />
+                </svg>
+              </button>
+              <img
+                src={item.productPicture[0]}
+                className="w-[35%] h-full object-cover rounded-sm"
+              ></img>
+              <div className="ml-2">
+                <h1 className="text-2xl ">{item.productName}</h1>
+                <h2 className="mt-2">{item.productNewPrice}$</h2>
+              </div>
+              <div className="h-full absolute right-2 ">
+                <div className=" h-full fl2 ">
+                  <button>
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      viewBox="0 0 24 24"
+                      fill="currentColor"
+                      class="w-5 h-5"
+                      onClick={() => {
+                        let newArray = [...quantity];
+                        let productPrice = item.productNewPrice;
+                        let newNumber = newArray[index] + 1;
+                        newArray.splice(index, 1);
+                        newArray.splice(index, 0, newNumber);
+                        setQuantity(newArray);
+
+                        setTotal((prev) => prev + productPrice);
+                      }}
+                    >
+                      <path
+                        fill-rule="evenodd"
+                        d="M11.47 7.72a.75.75 0 011.06 0l7.5 7.5a.75.75 0 11-1.06 1.06L12 9.31l-6.97 6.97a.75.75 0 01-1.06-1.06l7.5-7.5z"
+                        clip-rule="evenodd"
+                      />
+                    </svg>
+                  </button>
+                  <h1 className="text-center">{quantity && quantity[index]}</h1>
+                  <button>
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      viewBox="0 0 24 24"
+                      fill="currentColor"
+                      class="w-5 h-5"
+                      onClick={() => {
+                        if (quantity[index] === 1) {
+                          setSelectedProduct(item._id);
+                        } else {
+                          // onclick povecanja kolicine proizvoda
                           let newArray = [...quantity];
-                          let productPrice = item.productNewPrice;
-                          let newNumber = newArray[index] + 1;
+                          let newNumber = newArray[index] - 1;
                           newArray.splice(index, 1);
                           newArray.splice(index, 0, newNumber);
                           setQuantity(newArray);
-
-                          setTotal((prev) => prev + productPrice);
-                        }}
-                      >
-                        <path
-                          fill-rule="evenodd"
-                          d="M11.47 7.72a.75.75 0 011.06 0l7.5 7.5a.75.75 0 11-1.06 1.06L12 9.31l-6.97 6.97a.75.75 0 01-1.06-1.06l7.5-7.5z"
-                          clip-rule="evenodd"
-                        />
-                      </svg>
-                    </button>
-                    <h1 className="text-center">
-                      {quantity && quantity[index]}
-                    </h1>
-                    <button>
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        viewBox="0 0 24 24"
-                        fill="currentColor"
-                        class="w-5 h-5"
-                        onClick={() => {
-                          if (quantity[index] === 1) {
-                            setSelectedProduct(item._id);
-                          } else {
-                            // onclick povecanja kolicine proizvoda
-                            let newArray = [...quantity];
-                            let newNumber = newArray[index] - 1;
-                            newArray.splice(index, 1);
-                            newArray.splice(index, 0, newNumber);
-                            setQuantity(newArray);
-                            // zbrajanje totala
-                            let productPrice = item.productNewPrice;
-                            setTotal((prev) => prev - productPrice);
-                          }
-                        }}
-                      >
-                        <path
-                          fill-rule="evenodd"
-                          d="M12.53 16.28a.75.75 0 01-1.06 0l-7.5-7.5a.75.75 0 011.06-1.06L12 14.69l6.97-6.97a.75.75 0 111.06 1.06l-7.5 7.5z"
-                          clip-rule="evenodd"
-                        />
-                      </svg>
-                    </button>
-                  </div>
+                          // zbrajanje totala
+                          let productPrice = item.productNewPrice;
+                          setTotal((prev) => prev - productPrice);
+                        }
+                      }}
+                    >
+                      <path
+                        fill-rule="evenodd"
+                        d="M12.53 16.28a.75.75 0 01-1.06 0l-7.5-7.5a.75.75 0 011.06-1.06L12 14.69l6.97-6.97a.75.75 0 111.06 1.06l-7.5 7.5z"
+                        clip-rule="evenodd"
+                      />
+                    </svg>
+                  </button>
                 </div>
               </div>
-            );
-          })}
+            </div>
+          );
+        })}
       </div>
       <div className="p-2 h-[4.5%] flex relative items-center ">
         {" "}
         <h1 className="text-xl">Total: {total}$</h1>
         <button
-          onClick={handleBuyNow}
+          onClick={handleBoughtProducts}
           className="absolute right-0 bg-orange-500 text-white h-[80%] rounded-md w-[20%] hover:scale-95 "
         >
           Checkout
